Updated the property name authors to owners.
The use cased for contact information showed that this issue was
about contacting owners who could be a person, group, or org.

Authors is a harder term to track, expecially in open source, where
there are contributors, copyright holders, and so forth. Owners
is a better name as it is the person who manages or otherwise
controls the codebase.